+++ library/src/main/java/com/getstream/sdk/chat/adapter/MessageListItemViewHolder.java
@@ -284,6 +284,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder {
 
         if (isDeletedOrFailedMessage()
                 || message == null
+                || channelState.getLastMessage() == null
                 || TextUtils.isEmpty(message.getId())
                 || !messageListItem.getPositions().contains(MessageViewHolderFactory.Position.BOTTOM)
                 || !messageListItem.getMessageReadBy().isEmpty()