+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix Sent message progress infinite
 - Fix channel delete event handling in ChannelList view model
 - Fix attachment duplicated issue when message edit
+- Remove `UpdateChannelRequest`
 
 ## November 5th, 2019 - 3.2.1