+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix Concurrent modification when removing member from channel
 - Fix Edit/Delete thread parent message
+- Fix automention input issue
 - Fix Sent message progress infinite
 - Fix channel delete event handling in ChannelList view model
 - Fix attachment duplicated issue when message edit